+++ app/src/main/java/com/wildplot/android/rendering/XAxisHistoGram.java
@@ -159,11 +159,11 @@ public class XAxisHistoGram implements Drawable {
 	 * @param size specific size (width) of this bar
 	 */
 	private void drawBar(double x, double heigth, Graphics g, Rectangle field, double size) {
-		
-		
-		int[] pointUpLeft 		= plotSheet.toGraphicPoint(x, heigth, field);
-		int[] pointUpRight 		= plotSheet.toGraphicPoint(x+size, heigth, field);
-		int[] pointBottomLeft 	= plotSheet.toGraphicPoint(x, 0, field);
+
+
+        float[] pointUpLeft 		= plotSheet.toGraphicPoint(x, heigth, field);
+        float[] pointUpRight 		= plotSheet.toGraphicPoint(x+size, heigth, field);
+        float[] pointBottomLeft 	= plotSheet.toGraphicPoint(x, 0, field);
 		
 		if(heigth < 0) {
 			pointUpLeft 		= plotSheet.toGraphicPoint(x, 0, field);
